pubsub emulator rule moved to a separate project
