+++ spring-cloud-gcp-dependencies/pom.xml
@@ -191,6 +191,12 @@
 				<version>${project.version}</version>
 			</dependency>
 
+			<dependency>
+				<groupId>org.springframework.cloud</groupId>
+				<artifactId>spring-cloud-gcp-test-support</artifactId>
+				<version>${project.version}</version>
+			</dependency>
+
 			<!-- spring-cloud-gcp-starter-sql -->
 			<dependency>
 				<groupId>com.google.cloud.sql</groupId>
