+++ spring-cloud-gcp-pubsub-stream-binder/pom.xml
@@ -42,5 +42,11 @@
 			<version>3.1.0.BUILD-SNAPSHOT</version>
 			<scope>test</scope>
 		</dependency>
+		<dependency>
+			<groupId>org.springframework.cloud</groupId>
+			<artifactId>spring-cloud-gcp-test-support</artifactId>
+			<version>${project.version}</version>
+			<scope>test</scope>
+		</dependency>
 	</dependencies>
 </project>
